requirements.txt
gpiozero==1.6.2
pyusb==1.2.1
pigpio==1.78